+++ src/main/java/cn/nukkit/utils/VarInt.java
@@ -6,7 +6,7 @@ import java.io.OutputStream;
 
 /**
  * Tool class for VarInt or VarLong operations.
- * 
+ * <p>
  * Some code from http://wiki.vg/Protocol.
  *
  * @author MagicDroidX
@@ -24,7 +24,7 @@ public final class VarInt {
      */
     public static long encodeZigZag32(int v) {
         // Note:  the right-shift must be arithmetic
-        return ((v << 1) ^ (v >> 31)) & 0xFFFFFFFFL;
+        return (((long) v << 1) ^ (v >> 31)) & 0xFFFFFFFFL;
     }
 
     /**
